trim down itemviews
